Fix some SLA tests to use lowerTester as initiator
Fix some SLA tests that used upperTester as initiator, to
use the lowerTester. The issue was noticed when running
two different build configured hci application as IUT and
TST, wherein IUT was a peripheral only controller and TST
was full feature HCI controller. In this case, using the
upperTester as initiator failed, which is incorrect usage
in these tests.

Signed-off-by: Vinayak Kariappa Chettimada <vich@nordicsemi.no>