Handle fragmentation of ISO SDUs over HCI for transmission
Handle fragmentation of ISO SDUs over HCI by splitting data according
to the available buffer size and waiting for the respective number of
packets to be completed. Includes the addition of the LE Read Buffer
Size [v2] command and a minor fix to ensure that the SDU size does not
exceed the PDU size for some test cases.

Signed-off-by: Wolfgang Puffitsch <wopu@demant.com>