bsim transport: Use WAIT_WRESP insteado of WAIT
Use the new WAIT_WRESP command to the EDTT bridge
instead of the November modified WAIT when expecting
a reply for a wait command.

The change in the WAIT command functionality
broke backwards compatibility with the EDTT bridge
(and backwards compatibility from the newer bridge to
older EDTT version).
The bridge has now been fixed to avoid the backwards
compatibility issue with older (prior to Nov/23,
18744c937131104797ea0abf0dcfc84855a12311
) versions of the EDTT, but this means the WAIT
command no longer responds. Instead we need to use the new
WAIT_WRESP command.

Also check the response, if we get a UNKNOWN_COMMAND
(or anything !=0), we warn the user about the too old
EDTT instead of crashing badly.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>